+++ forgerock-auth/src/test/java/org/forgerock/android/auth/SSOBroadcastReceiverIntegrationTests.kt
@@ -12,6 +12,7 @@ package org.forgerock.android.auth
 
 import android.content.Context
 import android.content.Intent
+import android.net.Uri
 import com.squareup.okhttp.mockwebserver.Dispatcher
 import com.squareup.okhttp.mockwebserver.MockResponse
 import com.squareup.okhttp.mockwebserver.RecordedRequest
@@ -46,17 +47,7 @@ class SSOBroadcastReceiverIntegrationTests: BaseTest() {
         enqueue("/authTreeMockTest_Authenticate_NameCallback.json", HttpURLConnection.HTTP_OK)
         enqueue("/authTreeMockTest_Authenticate_PasswordCallback.json", HttpURLConnection.HTTP_OK)
         enqueue("/authTreeMockTest_Authenticate_success.json", HttpURLConnection.HTTP_OK)
-        server.enqueue(
-            MockResponse()
-                .addHeader(
-                    "Location",
-                    "http://www.example.com:8080/callback?code=PmxwECH3mBobKuPEtPmq6Xorgzo&iss=http://openam.example.com:8080/openam/oauth2&state=abc123&client_id=andy_app"
-                )
-                .setResponseCode(HttpURLConnection.HTTP_MOVED_TEMP)
-        )
-        enqueue("/authTreeMockTest_Authenticate_accessToken.json", HttpURLConnection.HTTP_OK)
-        enqueue("/userinfo_success.json", HttpURLConnection.HTTP_OK)
-        Config.getInstance().sharedPreferences =
+       Config.getInstance().sharedPreferences =
             context.getSharedPreferences(
                 defaultTokenManagerTest,
                 Context.MODE_PRIVATE
@@ -84,6 +75,21 @@ class SSOBroadcastReceiverIntegrationTests: BaseTest() {
                 }
             }
         FRUser.login(context, nodeListenerFuture)
+        server.takeRequest()
+        server.takeRequest()
+        server.takeRequest()
+
+        val request = server.takeRequest();
+
+        val state = Uri.parse(request.getPath()).getQueryParameter("state");
+        server.enqueue(MockResponse()
+            .addHeader("Location", "http://www.example.com:8080/callback?code=PmxwECH3mBobKuPEtPmq6Xorgzo&iss=http://openam.example.com:8080/openam/oauth2&" +
+                    "state=" + state + "&client_id=andy_app")
+            .setResponseCode(HttpURLConnection.HTTP_MOVED_TEMP));
+
+        enqueue("/authTreeMockTest_Authenticate_accessToken.json", HttpURLConnection.HTTP_OK)
+        enqueue("/userinfo_success.json", HttpURLConnection.HTTP_OK)
+
         assertNotNull(nodeListenerFuture.get())
         assertNotNull(FRUser.getCurrentUser())
         val future = FRListenerFuture<UserInfo>()