+++ forgerock-auth/src/test/java/org/forgerock/android/auth/devicebind/DeviceVerifierTests.kt
@@ -1,0 +1,11 @@
+package org.forgerock.android.auth.devicebind
+
+import org.forgerock.android.auth.callback.DeviceSigningVerifierCallback
+import org.junit.Test
+
+class DeviceVerifierTests {
+    @Test
+    fun testFoo() {
+       val testObject = DeviceSigningVerifierCallback()
+    }
+}